+++ route/evermist_island/__init__.py
@@ -1,4 +1,4 @@
-from route.evermist_island.intro import EvermistIsland
+from route.evermist_island.evermist_island import EvermistIsland
 
 __all__ = [
     "EvermistIsland",
